make index file just export, add decimals functions and bignumber.js package
