+++ package.json
@@ -25,6 +25,7 @@
     "@vegaprotocol/vegawallet-service-api-client": "^0.4.8",
     "apollo": "^2.33.9",
     "autoprefixer": "^10.4.2",
+    "bignumber.js": "^9.0.2",
     "classnames": "^2.3.1",
     "core-js": "^3.6.5",
     "date-fns": "^2.28.0",
